+++ crates/snforge-scarb-plugin/src/attributes/test_case.rs
@@ -38,11 +38,16 @@ fn test_case_internal(
     args: Arguments,
     _warns: &mut Vec<Diagnostic>,
 ) -> Result<TokenStream, Diagnostics> {
+    let named_args = args.named();
+    named_args.allow_only::<TestCaseCollector>(&["name"])?;
+
     let unnamed_args = args.unnamed();
     ensure_params_valid(func, &args.unnamed(), db)?;
 
     let func_name = func.declaration(db).name(db);
-    let case_fn_name = test_case_name(&func_name.text(db), &args, args_db)?;
+
+    let name_arg = named_args.as_once_optional("name")?;
+    let case_fn_name = test_case_name(&func_name.text(db), name_arg, &unnamed_args, args_db)?;
     let filtered_fn_attrs = collect_preserved_attributes_for_test_case(func, db);
 
     let signature = func.declaration(db).signature(db).as_syntax_node();